feat(#359): embed published url code
